Edit: Detail 구조 변경
